# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A movie database API with authentication, rate limiting, and caching\nThis API allows you to manage movies with CRUD operations",
        "title": "Movie Database API",
        "contact": {
            "name": "API Support",
            "email": "support@moviedb.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/watch-history": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get watch history for the authenticated user with pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "watch-history"
                ],
                "summary": "Get watch history",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default: 20, max: 100)",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by movie ID",
                        "name": "movie_id",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter by completion status",
                        "name": "completed",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Watch history retrieved successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Record or update watch history for a movie. Authenticated users only.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "watch-history"
                ],
                "summary": "Record watch history",
                "parameters": [
                    {
                        "description": "Watch history data",
                        "name": "watch",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/watchhistory.RecordWatchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Watch history recorded successfully",
                        "schema": {
                            "$ref": "#/definitions/watchhistory.RecordWatchResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Movie not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/watch-history/clear": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete all watch history entries for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "watch-history"
                ],
                "summary": "Clear all watch history",
                "responses": {
                    "200": {
                        "description": "Watch history cleared successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/watch-history/continue-watching": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get list of movies that the user is currently watching (not completed)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "watch-history"
                ],
                "summary": "Get continue watching",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit number of results (default: 10, max: 50)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Continue watching list retrieved successfully",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ContinueWatching"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/watch-history/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get aggregated watch statistics for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "watch-history"
                ],
                "summary": "Get watch statistics",
                "responses": {
                    "200": {
                        "description": "Watch statistics retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/models.WatchStats"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/watch-history/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a specific watch history entry by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "watch-history"
                ],
                "summary": "Delete watch history entry",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Watch History ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Watch history deleted successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - not your watch history",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Watch history not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.ContinueWatching": {
            "type": "object",
            "properties": {
                "last_position": {
                    "type": "integer",
                    "example": 2700
                },
                "movie": {
                    "$ref": "#/definitions/models.Movie"
                },
                "movie_id": {
                    "type": "integer",
                    "example": 1
                },
                "progress": {
                    "type": "number",
                    "example": 75.5
                },
                "watch_history_id": {
                    "type": "integer",
                    "example": 1
                },
                "watched_at": {
                    "type": "string",
                    "example": "2024-01-01T12:00:00Z"
                }
            }
        },
...
        "models.WatchStats": {
            "type": "object",
            "properties": {
                "average_progress": {
                    "type": "number",
                    "example": 82.5
                },
                "completed_movies": {
                    "type": "integer",
                    "example": 35
                },
                "favorite_genres": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "Action",
                        "Drama"
                    ]
                },
                "in_progress_movies": {
                    "type": "integer",
                    "example": 7
                },
                "last_watched_at": {
                    "type": "string",
                    "example": "2024-01-01T12:00:00Z"
                },
                "total_movies_watched": {
                    "type": "integer",
                    "example": 42
                },
                "total_watch_time": {
                    "description": "Total seconds",
                    "type": "integer",
                    "example": 151200
                }
            }
        },
...
        "watchhistory.RecordWatchRequest": {
            "type": "object",
            "required": [
                "last_position",
                "movie_id",
                "progress",
                "watch_duration"
            ],
            "properties": {
                "completed": {
                    "type": "boolean",
                    "example": false
                },
                "device_type": {
                    "type": "string",
                    "example": "web"
                },
                "last_position": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 2700
                },
                "movie_id": {
                    "type": "integer",
                    "example": 1
                },
                "progress": {
                    "type": "number",
                    "maximum": 100,
                    "minimum": 0,
                    "example": 75.5
                },
                "quality": {
                    "type": "string",
                    "example": "1080p"
                },
                "watch_duration": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 3600
                }
            }
        },
        "watchhistory.RecordWatchResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Watch history recorded successfully"
                },
                "watch_history": {
                    "$ref": "#/definitions/models.WatchHistory"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}